services/discount-service/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
redis>=4.5.0
psycopg2-binary==2.9.9
python-dateutil>=2.8.2
dj-database-url==2.2.0
